Tee kokonaismassamäärästä NUMERIC

+++ tietokanta/src/main/resources/db/migration/V1_858__.sql
@@ -6,7 +6,7 @@ ALTER TABLE pot2_alusta ADD COLUMN murske INTEGER REFERENCES pot2_mk_urakan_murs
 ALTER TABLE pot2_alusta ADD COLUMN kasittelysyvyys INTEGER;
 ALTER TABLE pot2_alusta ADD COLUMN leveys INTEGER;
 ALTER TABLE pot2_alusta ADD COLUMN pinta_ala INTEGER;
-ALTER TABLE pot2_alusta ADD COLUMN kokonaismassamaara INTEGER;
+ALTER TABLE pot2_alusta ADD COLUMN kokonaismassamaara NUMERIC;
 ALTER TABLE pot2_alusta ADD COLUMN massa INTEGER REFERENCES pot2_mk_urakan_massa (id);
 ALTER TABLE pot2_alusta ADD COLUMN sideaine INTEGER REFERENCES  pot2_mk_sideainetyyppi (koodi);
 ALTER TABLE pot2_alusta ADD COLUMN sideainepitoisuus NUMERIC(10, 1);
